fix(metaengine): duckdb float precision + persistent-DB test isolation
Follow-ups to the planned-table work, caught by the full verify gate and a
second live run:

- Reverted the float64 plan-type mapping back to "DOUBLE" (REAL broke
  duckdb's exact float8 round-trip: pi came back as 3.1415927) and instead
  taught the pgengine and mysqlengine planned-DDL translators to accept
  REAL/DOUBLE/FLOAT uniformly, mapping them to DOUBLE PRECISION (PG) and
  DOUBLE (MySQL). DuckDB gets its exact DOUBLE; pg/mysql no longer
  degenerate float columns to TEXT. LayoutPlanner CGo tests green again.
- The MariaDB planned-table tests now drop their planned tables and
  meta_map rows before AND after each test: cqrs_test persists between
  runs, and the first pass left rows that made the second run's
  nil-prev MapUpdate assertion fail. Two consecutive live runs are green.

Gates: per-module lint 0 issues both engines, check-duplication green,
duckdb CGo layout suite green, PG + MariaDB planned suites green.

+++ metaengine/layout_type_test.go
@@ -29,7 +29,7 @@ func TestBuildLayoutPlanFromType_InfersColumnTypes(t *testing.T) {
 	cases := map[string]string{
 		"status":   "TEXT",    // string
 		"priority": "INTEGER", // int
-		"score":    "REAL",    // float64 — canonical float name: pg maps REAL→DOUBLE PRECISION, mysql REAL→DOUBLE, duckdb accepts REAL
+		"score":    "DOUBLE", // float64 — duckdb-exact; pg/mysql translators map DOUBLE → float8
 		"name":     "TEXT",    // string
 	}
